// src/app/_dto/TblHTabDispHide.dto.ts
import { TbldTabDispHideCo } from "./TbldTabDispHideCo.dto";

export class TblHTabDispHide {

    HTabDispHide_SysID: number;
    HTabDispHide_Code: string;
    HTabDispHide_Name: string;
    HTabDispHide_AcDe_Yn: string ="N";
    HTabDispHide_AcDe_SysID: number;
    HTabDispHide_AcDe_Code: string;
    HTabDispHide_AcDe_Name: string;
    TbldTabDispHideCo:TbldTabDispHideCo[]=[]

}